criação do metodo update de Aluno

+++ src/main/java/org/example/escola/Repository/AlunoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AlunoRepository extends CrudRepository<Aluno, Integer> {
+public interface AlunoRepository extends CrudRepository<Aluno, Long> {
 }
